start to refactor plugins to use a more extensible model. Day Night is the first to get the treatment
