+++ Assets/Digital Painting/Docs/CreatingAScene.md
@@ -17,6 +17,8 @@ Unity Pro and some other assets.
 
 Note, that your scene must have a terrain within it. This is used for agent navigation.
 
+FIXME: Validate the scene has a terrain when the DigitalPaintingManager is added
+
 The Meadow environment 
 asset recommends you to setup some of the lighting and camera effects, but it should be
 noted that some plugins for THe Digital Painting also advise specific lighting and camera 
@@ -42,9 +44,9 @@ defaults this is pretty easy to do. Firstly configure the pacakges to be install
     "com.unity.cinemachine": "2.2.8",
 ```
 
-FIXME: When we have made releases of the other assets we use install those via the package manager. For now see the 'From Source' section below.
+Next you need to insert the Digital Painting code.
 
-Next you need to insert the Digital Paintin code. You can either work from the source or from a released package:
+FIXME: When we have made releases of the other assets we use install those via the package manager. For now see the 'From Source' section below.
 
 ### From Source
 
@@ -56,12 +58,12 @@ git submodule add git@github.com:DigitalPainting/DigitalPainting.git
 git submodule add git@github.com:DigitalPainting/Flying-Pathfinding.git
 ```
 
-Note there will be some warnings when the code is imported. This is caused by a limitation of the way Unity imports packages that are git submodules. So far, apart from the annoying warnings on import and startup we've not found any problems.
+Note there may be some warnings when the code is imported. This is caused by a limitation of the way Unity imports packages that are git submodules. So far, apart from the annoying warnings on import and startup we've not found any problems.
 
 ## Configure Your Scene to use the Digital Painting
 
-  * Add the `DigitalPaintingManager` prefab to your scene
-  * Add the Octree prefab to the scene and ensure the Box Collider covers your terrain (for a standard 500 x 500 terrain this means you should position the octree at 250, 0, 250 and set the box collider scale to 500, 100, 500 - note that the y value changes how high your flying agents can go)
+  * `Window > Wizards Code > Open Digital Painting Manager`
+  * Select the `Standard` tab and clidk `Add Digital Painting Manager`
   * Add an agent to the scene - the easiest way to get started is to add the Fairy agent that comes with the Digital Painting package, see the 'Spawning Agents' section of the [agent](agents.md) documentation.
 
 ## DONE!
