+++ Assets/Digital Painting/Docs/Development.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e48e0dc9bd39ee947bf4ae0fa717a0bc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
