+++ Assets/Digital Painting/Docs/Development/EditorConfig.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ff5cd9641aed54240a68f4fbf6ea7d83
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
