--- Assets/Digital Painting/Plugins/DayNightCycle_WeatherMaker.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: db2ea220e0e95704a9a86501293516c0
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
