+++ Assets/Digital Painting/ScriptableObjects/SimpleDayNightProfileSO.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d6d234b33cb227c4788868c04885ae87, type: 3}
+  m_Name: SimpleDayNightProfileSO
+  m_EditorClassIdentifier: 
+  dawnDuskFog: {r: 0.5943396, g: 0.47457296, b: 0.3504361, a: 1}
+  dayFog: {r: 0.8243592, g: 0.8937482, b: 0.8962264, a: 1}
+  nightFog: {r: 0.07186722, g: 0.075944856, b: 0.1792453, a: 1}
