+++ Assets/Digital Painting/ScriptableObjects/SimpleDayNightProfileSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 96a5d14ec4ca7c3488ce2ba2ca207b7d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
