+++ Assets/Digital Painting/Scripts/Editor/EditorConfigScriptableObject.cs
@@ -6,11 +6,17 @@ using wizardscode.digitalpainting;
 
 public class EditorConfigScriptableObject : ScriptableObject
 {
-    public static string LatestVersion = "0.0.11";
+    public static string LatestVersion = "0.0.18";
+        
+    [Header("Prefabs")]
     public DigitalPaintingManager ManagerPrefab;
-    public string version;
-    public string DocsIndexURL = "https://github.com/DigitalPainting/DigitalPainting/tree/master/Assets/Digital%20Painting/Docs";
+    public Octree FlyingPathfinderPrefab;
+
+    [Header("Informational")]
     public string ManagerName = "Digital Painting Manager";
+    public string DocsIndexURL = "https://github.com/DigitalPainting/DigitalPainting/tree/master/Assets/Digital%20Painting/Docs";
+    
+    internal string version;
 
     public void Init()
     {
@@ -26,6 +32,7 @@ public class EditorConfigScriptableObject : ScriptableObject
         Init();
         Upgrade(oldConfig, "ManagerName");
         Upgrade(oldConfig, "ManagerPrefab");
+        Upgrade(oldConfig, "FlyingPathfinderPrefab");
     }
 
     private void Upgrade(EditorConfigScriptableObject old, String propertyName)
