+++ Assets/Digital Painting/Scripts/Environment/AbstractDayNightProfile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2bdab2b9290857a4991e05fee8d8464c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
