--- Assets/Digital Painting/Scripts/Environment/SimpleDayNightCycle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d6d234b33cb227c4788868c04885ae87
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
