+++ Assets/Digital Painting/Scripts/Environment/SimpleDayNightProfile.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0fec34d1e882abe4dac25e3d6a025fea
+guid: d6d234b33cb227c4788868c04885ae87
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
