--- Assets/Digital Painting/Scripts/Plugin/WeatherMakerDayNightPlugin.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class WeatherMakerDayNightPlugin : AbstractDayNightPluginDefinition
-{
-    public override string GetPluginImplementationClassName()
-    {
-        throw new System.NotImplementedException();
-    }
-
-    public override string GetReadableName()
-    {
-        throw new System.NotImplementedException();
-    }
-
-    public override string GetURL()
-    {
-        throw new System.NotImplementedException();
-    }
-}
