+++ Assets/Digital Painting/Scripts/Plugin/WeatherMakerDayNightPluginDefinition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cff5af3938977f54080294ac1d62dfcd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
